+++ src/test/kotlin/io/openfuture/chain/rpc/controller/transaction/DelegateTransactionControllerTests.kt
@@ -1,6 +1,9 @@
 package io.openfuture.chain.rpc.controller.transaction
 
 import io.openfuture.chain.config.ControllerTests
+import io.openfuture.chain.core.model.entity.block.MainBlock
+import io.openfuture.chain.core.model.entity.block.payload.MainBlockPayload
+import io.openfuture.chain.core.model.entity.transaction.confirmed.DelegateTransaction
 import io.openfuture.chain.core.model.entity.transaction.payload.DelegateTransactionPayload
 import io.openfuture.chain.core.model.entity.transaction.unconfirmed.UnconfirmedDelegateTransaction
 import io.openfuture.chain.core.service.DelegateTransactionService
@@ -20,6 +23,10 @@ class DelegateTransactionControllerTests : ControllerTests() {
     @MockBean
     private lateinit var service: DelegateTransactionService
 
+    companion object {
+        private const val DELEGATE_TRANSACTION_URL = "/rpc/transactions/delegate"
+    }
+
 
     @Test
     fun addTransactionShouldReturnAddedTransaction() {
@@ -31,7 +38,7 @@ class DelegateTransactionControllerTests : ControllerTests() {
 
         given(service.add(transactionRequest)).willReturn(unconfirmedDelegateTransaction)
 
-        val actualResponse = webClient.post().uri("/rpc/transactions/delegates")
+        val actualResponse = webClient.post().uri(DELEGATE_TRANSACTION_URL)
             .body(Mono.just(transactionRequest), DelegateTransactionRequest::class.java)
             .exchange()
             .expectStatus().isOk
@@ -39,6 +46,24 @@ class DelegateTransactionControllerTests : ControllerTests() {
             .returnResult().responseBody!!
 
         assertThat(actualResponse).isEqualToComparingFieldByField(expectedResponse)
+    }
+
+    @Test
+    fun getTransactionByHashShouldReturnTransaction() {
+        val hash = "hash"
+        val mainBlock = MainBlock(1, 1, "previousHash", 1, "hash", "signature", "publicKey", MainBlockPayload("merkleHash")).apply { id = 1 }
+        val expectedTransaction = DelegateTransaction(1L, 1L, "senderAddress", "hash",
+            "senderSignature", "senderPublicKey", mainBlock, DelegateTransactionPayload("delegateKey")).apply { id = 1 }
+
+        given(service.getByHash(hash)).willReturn(expectedTransaction)
+
+        val actualTransaction = webClient.get().uri("$DELEGATE_TRANSACTION_URL/$hash")
+            .exchange()
+            .expectStatus().isOk
+            .expectBody(DelegateTransaction::class.java)
+            .returnResult().responseBody!!
+
+        assertThat(actualTransaction).isEqualTo(expectedTransaction)
     }
 
 }