+++ src/test/kotlin/io/openfuture/chain/rpc/controller/transaction/VoteTransactionControllerTests.kt
@@ -1,6 +1,9 @@
 package io.openfuture.chain.rpc.controller.transaction
 
 import io.openfuture.chain.config.ControllerTests
+import io.openfuture.chain.core.model.entity.block.MainBlock
+import io.openfuture.chain.core.model.entity.block.payload.MainBlockPayload
+import io.openfuture.chain.core.model.entity.transaction.confirmed.VoteTransaction
 import io.openfuture.chain.core.model.entity.transaction.unconfirmed.UnconfirmedVoteTransaction
 import io.openfuture.chain.core.model.entity.transaction.vote.VoteTransactionPayload
 import io.openfuture.chain.core.service.VoteTransactionService
@@ -19,6 +22,10 @@ class VoteTransactionControllerTests : ControllerTests() {
     @MockBean
     private lateinit var service: VoteTransactionService
 
+    companion object {
+        private const val VOTE_TRANSACTION_URL = "/rpc/transactions/vote"
+    }
+
 
     @Test
     fun addTransactionShouldReturnAddedTransaction() {
@@ -30,7 +37,7 @@ class VoteTransactionControllerTests : ControllerTests() {
 
         given(service.add(transactionRequest)).willReturn(unconfirmedVoteTransaction)
 
-        val actualResponse = webClient.post().uri("/rpc/transactions/votes")
+        val actualResponse = webClient.post().uri(VOTE_TRANSACTION_URL)
             .body(Mono.just(transactionRequest), VoteTransactionRequest::class.java)
             .exchange()
             .expectStatus().isOk
@@ -38,6 +45,24 @@ class VoteTransactionControllerTests : ControllerTests() {
             .returnResult().responseBody!!
 
         assertThat(actualResponse).isEqualToComparingFieldByField(expectedResponse)
+    }
+
+    @Test
+    fun getTransactionByHashShouldReturnTransaction() {
+        val hash = "hash"
+        val mainBlock = MainBlock(1, 1, "previousHash", 1, "hash", "signature", "publicKey", MainBlockPayload("merkleHash")).apply { id = 1 }
+        val expectedTransaction = VoteTransaction(1L, 1L, "senderAddress", "hash",
+            "senderSignature", "senderPublicKey", mainBlock, VoteTransactionPayload(1, "delegateKey")).apply { id = 1 }
+
+        given(service.getByHash(hash)).willReturn(expectedTransaction)
+
+        val actualTransaction = webClient.get().uri("$VOTE_TRANSACTION_URL/$hash")
+            .exchange()
+            .expectStatus().isOk
+            .expectBody(VoteTransaction::class.java)
+            .returnResult().responseBody!!
+
+        assertThat(actualTransaction).isEqualTo(expectedTransaction)
     }
 
 }