+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - RPC: Get explorer info
 - RPC: Get list of blocks 
 - RPC: Get list of transactions 
+- Test to new controller methods
 
 ## [0.5.0] - 2018-08-10
 ### Added